rltiles: Allow referencing generated images by enum
Wherever image filenames are used, you can now also refer to an
already-generated tile by its enum name. The first such matching tile is
chosen; there is no support (yet?) for selecting a variation.

The motivation for this feature is to reduce duplication and ensure that
changes between tiles are synchronized.