Positional magic ❄️: New Ice spell: Hailstorm
Hailstorm is a Level 4 Conjuration/Ice spell. It replaces Throw Icicle
in the IE starting book. When cast it fires a ring of inner radius 2 and
outer radius 3 of BEAM_ICE (like throw icicle) flavoured beams. Icy monsters
are immune, so this spell gets along with ice beasts and simulacra but
not other minions. Having it skip allies all together is something to
consider. Currently, Veh does not add a fourth ring, that would almost
double the number of affected squares.

The damage is that of throw icicle but the accuracy is higher.

[Thanks to cwz testers and Bacchus for early feedback on this spell.]