Add widget focus cycling
This implements a more versatile input focus layer for local tiles and
console. Widgets can now be focusable or not, depending on derived
type. On UI hierarchy change, a DFS is performed of all widgets and all
focusable widgets are collected into an array, which is then used for
cycling between focusable elements.

This commit adjusts keyboard event delivery. The keyboard event (key up
or down) is delivered first to the focused widget, which has a chance to
handle the event and stop propagation. Next, tab, shift+tab, and esc are
handled; if one of these keys is pressed, propagation stops. Otherwise,
propagation continues upwards.

Input focus on webtiles is decoupled from input focus on the server.
The current plan is that the webtiles implementation for a given UI
layout will have corresponding focusable HTML elements, and that UI
state, including focus, will be synchronized in both directions.