Remove/mark unused parameters
This commit transitively removes unused parameter from all over crawl.
If the unused parameters look like they're part of an interface, I've
left them as is, but commented out the parameter name. In general, if
I'm less confident that a parameter is intentionally ignored, i've
marked it with UNUSED().

One pain point is that crawl has a lot of conditional compilation, and
variables are sometimes only used by one ifdef branch. I've still used
UNUSED in these cases, but it's less than ideal.