Rework a worm vault (again)
This commit reworks worms_lemuel to place more reasonable numbers of
monsters in the earliest depths and to scale monsters and loot up with
depth. For example, in D:7-10 it places only two on-average and 3 at
maximum. The moths of wrath and/or torpor snails are also scaled more
reasonably, only beginning to appear in Lair. Previously the vault used
absolute depth for monster placement, but now it looks at branch and
branch level. The vault's layout is slightly changed to accommodate the
potential max number of spawns.

I've also made the vault place within-branch monsters in Swamp/Spider,
since placing swamp worms in Spider is not too interesting. Now it uses
demonic crawlers in Spider and only torpor snails as the "assistant"
monster there, since crawlers can't be berserked by moth of wrath.