Fix menu filtering from crashing (damerell, 12120)
cursorto() moves the on-screen cursor to the correct position, given the
cursor offset in the input text. However, resumable_line_reader never
renders to the screen directly, so it's safe to NOP this function
entirely.

The crash itself seems to have something to do with the line reader
start position, but I'm less interested in fixing this hacky mess
because a replacement (textentry widget) is almost finished.