Don't feel safe near exploding monsters
Change i_feel_safe to alert the player instead of waiting near friendly
fulminant prisms, ball lightning, or ballistomyete spores. Currently
this is slightly overzealous and will refuse to wait in sight of a
fulminant prism regardless of distance, but this minor edge case is
preferable to players waiting after a fight with Fedhas or CBL and dying
after pressing 5.