Improve behavior of `remember_name = false`
There were several cases where this option wasn't working as expected.
Two main ones:

* If a name was present in the sticky prefs, but this option was false,
  the saved name would still be loaded.
* It was ignored on restart after a game end.

This commit also lets `*` work to pick a random name at the main menu,
not just at the name prompt. It's a bit of a hidden feature for now, not
sure how best to document it to users.