Fix redraw issue when game ends (Goratrix, 12124)
Fully redrawing the screen is only necessary when running, since a
more() may be issued while running and the map should be updated to
show the correct location. Otherwise, fully redrawing currently causes
redraw issues and crashes if more() forces a redraw during monster
moves; use the old draw code in this case for now.