Positional magic 🔥: New spell Foxfire
Foxfire is a L1 Conjurations/Fire spell that replaces Flame Tongue.
Foxfire conjures two foxfires adjacent to the caster which seek the
nearest target and burn the target, or dissapate if they fail to reach
the target after making 8 moves. Monsters and the player can swap with
foxfire, but doing so triggers the foxfire attack.

Each foxfire burn does slightly more than half the damage of Flame
Tongue, but does not miss. This makes the spell better than FT when
fighting in the open (discounting the delay), but much worse in
corridors.

Foxfires have see invisible; with their high move speed this allows them
to attack first, dissipating and allowing the monster to then move
towards the player.

[Thanks to hellmonk, tealizard, Doesnt, EN0N, and cwz testers for
 feedback in development of this spell.]