Adjust Trog weapon gifting (ebering)
Only gift melee weapons with a restricted set of brands (plain, vorpal,
flaming, antimagic). This is to differentiate Trog gifts from Okawaru's
further, and in line Trog's theme of crude anger, fire, and hatred of magic.
Gifting only melee weapons also makes Trog slightly simpler, which is a
bonus for new players.

This also slightly modifies the minimum enchantment of gifts from Trog
and Okawaru. Previously the minimum was 0 or 1 (random per-item), now it
is always 0.

[ Committer's note: Significantly revised the approach in the change to
  acquirement. The acquirement system did not previously distinguish
  ranged and melee weapons, doing so required a rather deep dive.
  Closes #1208 ]