Ely's Divine Protection always protects at *****
At present, the always protection of this power begins at 131 piety.
This figure is slightly different from when the player's faith becomes
***** (120). This subtle difference is confusing for a player who
doesn't read a spoiler because nothing in-game alerts the player to the
moment when the odds of protection are 100%, though the guarantee is
described in the ^ screen. This confusion can be resolved by aligning
the point at which the protection probability is 100% to the moment at
which the number of stars changes.

As a result players will be able to clearly see when this sentence means
without a spoiler: Especially pious individuals may be certain of this
help.

[ Committer's note: revised the commit message. Closes #1212 ]