Don't trap monsters in Zig pillars (AliasTheSpectator)
The Ziggurat pillar vault ziggurat_pillar_centre_drop places monsters
behind fully opaque walls, providing a teleport trap to release them.
This no longer works with LOS requirements for trap triggering, so make
the vault always have transparent rock or stone windows. I've also
modified the vault substitutions so that the windows will always be
diggable when the rest of the vault is.