Fix (?) a race condition with ui state syncing
When starting up the game and opening the seed selection dialog, the
game fills in a seed and attempts to sync that to the js client. If the
js client doesn't receive and process that message fast enough, it will
send a sync message to the game with a blank text field; in my local
setup these actually cross paths and things end up in an inconsistent
state: the game thinks the field is empty, but there is still a number
showing on the client side. (If the player does more things in the UI
that number will be resynced, so it is only on immediately pressing
enter that there is buggy behavior.)

This fix is very brute-force, and I'm not sure it'll really get all the
cases. I'm generally worried that this syncing mechanism can have
race-cases like this beyond just the game start, and also it seems like
the possibility of messages crossing paths might need to be dealt with
more consistently.