Make fedhas_protects less likely to crash if player is in the mix
This function was routinely called with *act->as_monster(), which is a
fairly bad paradigm since player::as_monster() returns nullptr, and so a
segfault was guaranteed of a protecting monster check is missed. As long
as this is so, it is better for this function to take a monster pointer
instead of reference.