Integrate catch2 testing framework into crawl.
In order to compile and run the catch2 tests, use the "make
catch2-tests" command.

If you want to see an example test file, look at
catch2-tests/test_english.cc. It should be possible to test any leaf
functions in crawl just by copying that file and making minor changes.

Future work:

For some reason whenever I try to change the names of the test files
from using underscores to using dashes, I get a "No such file or
directory" error for pty.h (util/fake_pty.c:5:12, which #includes
<pty.h>).  Note I am using the current way of compiling on Windows
suggested in INSTALL.txt. It would be good to fix whatever is causing
this so the test files can use dashes to separate words like most crawl
files.

Travis ci should be configured to verify the catch2 tests are working at
the same time it compiles. This should be pretty easy - it just needs to
run the "make catch2-tests" command and verify that command exits
without an error.

It would be good to eventually create a way of faking some globals to
enable easier testing - player you is a big offender.