Buff Leda's Liquefication
Make the caster immune to the maluses imposed by liquefied ground. This
goes all the way back to the introduction of the spell in 2010, and the
justification given then was simulationist.

Hexes are intended to support other combat styles, but the attack
fumbling and slow movement penalties meant that Leda's made little sense
for any archetype and as a result the spell was little-used.

The shrinking radius and re-cast delay imposed by the radius 0 period
should provide sufficient buffer against aggressive kiting strategies
with the spell; if this proves incorrect the movement penalty can be
restored (without the fumble penalty).