Refactor conducts for Zin's sanctuary
Before this commit, any code that applied god conducts had to make its
own additional checks for sanctuary violations and potentially call
remove_sanctuary(). This commit moves this handling into
set_attack_conduct() and the god_conduct_trigger class so that separate
sanctuary checks aren't necessary.

Technical note about Toxic Radiance and #999: A comment noted that the
god conduct had to be applied before any toxic radiance damage in order
to properly apply conducts for monsters killed the turn the spell is
cast, but this was not true as of the current code. A unique_ptr in the
trigger holds a monster object that has the properties of the original
monster copied by value before any damage is made; any victim data
necessary to apply the conduct is hence still available.