Add an aliveness check for LRDing a monster
Don't let LRD attempt to fragment using a dead monster as the source.
This probably could not arise in practice, as currently nothing can
trigger an LRD on a monster that has already died. Still, if the monster
at the target is dead, we should ensure that any appropriate terrain at
the location is used as the LRD source instead. Also ensure that no god
conducts are applied for LRD happening at a dead monster's location.