Allow pushing/pulling monsters directly into deep water/lava with Force Lance/Force Lasso. Properly credit the correct actor when a death is due to confusing/pushing/pulling into a hazard (Terrain or Cloud). Fix Force Lance/Force Lasso not working with Warped Staves.
This will give players XP for drowning kills. Note: You still don't get XP for kills via CREATED terrain (IE if a Chaos Cast made lava under something); I may or may not change this in the future. In the case of clouds, if they are natural (volcano, ice cave, etc.) blame will always go to the confuser; if something created them the blame is randomized.